+++ .meteor/packages
@@ -17,7 +17,6 @@ standard-minifier-js@1.2.1    # JS minifier run for production mode
 es5-shim@4.6.15                # ECMAScript 5 compatibility for older browsers.
 ecmascript@0.6.1              # Enable ECMAScript2015+ syntax in app code
 
-iron:router
 twbs:bootstrap
 mrt:jquery-validate
 themeteorchef:jquery-validation
@@ -33,3 +32,6 @@ reactive-dict@1.1.8
 shell-server
 chrismbeckett:toastr
 email
+kadira:flow-router
+kadira:blaze-layout
+arillo:flow-router-helpers
