+++ .meteor/versions
@@ -7,6 +7,7 @@ accounts-password@1.3.3
 accounts-ui@1.1.9
 accounts-ui-unstyled@1.1.12
 allow-deny@1.0.5
+arillo:flow-router-helpers@0.5.2
 autoupdate@1.2.11
 babel-compiler@6.13.0
 babel-runtime@1.0.1
@@ -38,6 +39,7 @@ cfs:upload-http@0.0.20
 cfs:worker@0.1.4
 check@1.2.4
 chrismbeckett:toastr@2.1.2_1
+coffeescript@1.0.17
 ddp@1.2.5
 ddp-client@1.2.9
 ddp-common@1.2.8
@@ -61,15 +63,9 @@ html-tools@1.0.11
 htmljs@1.0.11
 http@1.2.10
 id-map@1.0.9
-iron:controller@1.0.12
-iron:core@1.0.11
-iron:dynamic-template@1.0.12
-iron:layout@1.0.12
-iron:location@1.0.11
-iron:middleware-stack@1.1.0
-iron:router@1.0.13
-iron:url@1.0.11
 jquery@1.11.10
+kadira:blaze-layout@2.3.0
+kadira:flow-router@2.12.1
 launch-screen@1.0.12
 less@2.6.5
 livedata@1.0.18
@@ -124,3 +120,4 @@ underscore@1.0.10
 url@1.0.11
 webapp@1.3.12
 webapp-hashing@1.0.9
+zimme:active-route@2.3.2
