--- client/advisor/advisor.html
@@ -1,36 +1,0 @@
-<template name = "advisor">
-  {{>navbar}}
-  {{# if currentUser}}
-    <!--Navigation Bar Code including links to IDs-->
-  <div class="section" id="advisor">
-    <div class="container">
-      <div class="row">
-        <!--First Column of Advisor Section-->
-        <div class="col-sm-1">
-          <br>
-          <br>
-          <br>
-          <img src="images/GradPI-logo.png">
-        </div>
-        <!--Second Column of Advisor Section-->
-      </div>
-    </div>
-  </div>
-  <div class="container">
-    <div class="mix">
-      <a href={{linker}}>
-        <div class = "crop">
-          <img src={{find_advisor_info}} alt=''>
-        </div>
-        <h4><b>{{ thisAdvisor.name }} {{ thisAdvisor.school }}</b></h4>
-        <p></p>
-      </a>
-    </div>
-  </div>
-  {{ else}}
-      {{> about}}
-  {{/if}}
-</template>
-<!-- {{#each school}}
-<li><a href="/school/{{this._id}}">{{this.name}}</a></li>
-{{/each}} -->
