--- client/find-pi/find-pi.html
@@ -1,79 +1,0 @@
-<template name = "findPi">
-	<body>
-		<!--Navigation Bar Code including links to IDs-->
-		{{>navbar}}
-		{{# if currentUser}}
-		<div class="section" id="about">
-			<div class="container">
-				<div class="row">
-
-					<!--First Column of About Section-->
-					<div class="col-sm-1">
-						<br>
-						<br>
-						<br>
-						<img src="images/GradPI-logo.png">
-					</div>
-
-					<!--Second Column of About Section-->
-					<div class="col-sm-8 col-sm-offset-3">
-						<br>
-						<br>
-						<br>
-						<br>
-						<br>
-						<div class="form-group">
-    						<label for="School">School</label>
-							<select class="form-control" id="schoolSelect">
-							 	  <option class="school" value="">Select School</option>
-							      <option class="school" value="Albertus Magnus College">Albertus Magnus College</option>
-							      <option class="school" value="Columbia">Columbia University</option>
-							      <option class="school" value="Quinnipiac">Quinnipiac University</option>
-							      <option class="school" value="Southern Connecticut State University">Southern Connecticut State University</option>
-							      <option class="school" value="University of New Haven">University of New Haven</option>
-							      <option class="school" value="Yale">Yale University</option>
-							</select>
-						</div>
-
-					    <div class="form-group">
-    						<label for="Department">Department</label>
-							<select class="form-control" id="deptSelect">
-							      <option class="dept" value="">Select Department</option>
-							      <option class="dept" value="Bioinformatics">Bioinformatics</option>
-							      <option class="dept" value="Biomedical Sciences">Biomedical Sciences</option>
-							      <option class="dept" value="Engineering">Engineering</option>
-							      <option class="dept" value="Humanities">Humanities</option>
-							      <option class="dept" value="Public Health">Public Health</option>
-							</select>
-						</div>
-					</div>
-				</div>
-			</div>
-		</div>
-
-		<div id="Container" class="container">
-			{{#each findAdvisors }}
-				<div class="col-sm-3">
-					<div class="advisor">
-						<br>
-						<br>
-						{{name}}
-						<br>
-						{{school}}
-						<br>
-						{{dept}}
-						<br>
-						<div class="crop">
-							<img src={{image}}>
-						</div>
-						<br>
-						<br>
-					</div>
-				</div>
-			{{/each}}
-		</div> 
-		{{ else}}
-      		{{> about}}
-  		{{/if}}   							
-	</body>
-</template>
