--- client/find-school/find-school.html
@@ -1,49 +1,0 @@
-<template name = "schools">
-	<body>
-		<!--Navigation Bar Code including links to IDs-->
-		{{>navbar}}
-		{{#if currentUser}}
-			<div class="section" id="school">
-				<div class="container">
-					<div class="row">
-						<!--First Column of Find-School Section-->
-						<div class="col-sm-1">
-							<br>
-							<br>
-							<br>
-							<img src="images/GradPI-logo.png">
-						</div>
-						<!--Second Column of About Section-->
-						<div class="col-sm-5 col-sm-offset-3">
-							<p>
-								<br>
-								<br>
-								<br>
-								<br>
-								<b>Schools</b>
-								<br>
-								<ul>
-									{{#each school}}
-									<li><a href="/school/{{this._id}}">{{this.name}}</a></li>
-									{{/each}}
-
-									<!-- <li>Gateway Community College</li>
-									<li>Naugatuck Valley Community College</li>
-									<li>Paier College of Art</li>
-									<li>Post University</li>
-									<li>Quinnipiac University</li>
-									<li>Southern Connecticut University</li>
-									<li>University of New Haven</li>
-									<li>University of Connecticut</li>
-									<li>Yale University</li> -->
-								</ul>
-							</p>
-						</div>
-					</div>
-				</div>
-			</div>
-			{{ else}}
-      			{{> about}}
-  			{{/if}} 	
-	</body>
-</template>
