+++ client/layouts/mainLayout/mainLayout.html
@@ -1,0 +1,4 @@
+<template name= "mainLayout">
+	{{> navbar}}
+	{{> Template.dynamic template=content}}
+</template>
