+++ client/lib/redirector.js
@@ -1,0 +1,5 @@
+Tracker.autorun(function(){
+	if(!Meteor.userId()){
+		FlowRouter.redirect('/');
+	}
+});
