--- client/main.html
@@ -1,6 +1,0 @@
-<head>
-      <!--Inclusion of Bootstrap Content Delivery Networks for Font-Awesome, CSS, and JS (CDN)-->
-      <!-- <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.5.0/css/font-awesome.min.css">
-      <link type= "text/css" rel = "stylesheet" href = "stylesheet.css"/>
-      <title>Grad P.I.</title> -->
-</head>
