--- client/main.js
@@ -1,2 +1,0 @@
-import '../imports/startup/accounts-config.js';
-import '../imports/ui/body.js';
