+++ client/shared/rating-stars/rating-stars.js
