--- client/templates/choose-pi.html
@@ -1,37 +1,0 @@
-<template name = "choose-pi">
-	<body>
-		<!--Navigation Bar Code including links to IDs-->
-		{{>navbar}}
-			<div class="section" id="rate-pi">
-				<div class="container">
-					<div class="row">
-						<!--First Column of About Section-->
-						<div class="col-md-1">
-							<br>
-							<br>
-							<br>
-							<br>
-							<h1>PIs</h1>
-						</div>
-						<!--Second Column of About Section-->
-						<div class="col-md-5 col-md-offset-3">
-							<p>
-								<br>
-								<br>
-								<br>
-								<br>
-								<ul>
-									<li>Hongyu Zhao</li>
-									<li>Jeffrey Townsend</li>
-									<li>Yuval Kluger</li>
-									<li>Ronald Breaker</li>
-									<li>Craig Crews</li>
-									<li>Forrest Crawford</li>
-								</ul>
-							</p>
-						</div>
-					</div>
-				</div>
-			</div>	
-	</body>
-</template>
