--- client/templates/contact.html
@@ -1,32 +1,0 @@
-<template name = "contact">
-    {{>navbar}}
-      <!--Navigation Bar Code including links to IDs-->
-      <div class="section" id="contact">
-        <div class="container">
-          <div class="row">
-            <!--First Column of About Section-->
-            <div class="col-md-1">
-              <br>
-              <br>
-              <br>
-              <h1>GRAD PI</h1>
-            </div>
-            <!--Second Column of About Section-->
-            <div class="col-md-5 col-md-offset-3">
-              <p>
-                <br>
-                <br>
-                <br>
-                <br>
-                <b>CONTACT</b>
-                <br>
-                <br>
-                Perhaps the most important decision a student makes in graduate school is choosing an advisor or Principal Investigator (PI). This decision can make or break the student. However, there are few resources out there to reliably assess potential matches. It is essential that grad students know what they are getting into when they pick an advisor. Factors such as working autonomy, project direction, empathy, and impact in the field are all essential things to know about a PI and his/her lab.
-
-                Solution: GradPI will allow graduate students to evaluate their advisor in the same way that Rate my Professors permits undergrads to assess their college professors. This site will help to keep advisors accountable by allowing an open forum where grad students can evaluate them based on availability, support, management style, and other factors that are important to students' success.
-              </p>
-            </div>
-          </div>
-        </div>
-      </div>    
-</template>
