--- client/templates/layout.html
@@ -1,3 +1,0 @@
-<template name= "layout">
-	{{>yield}}
-</template>
