--- imports/startup/accounts-config.js
@@ -1,5 +1,0 @@
-import { Accounts } from 'meteor/accounts-base';
- 
-Accounts.ui.config({
-  passwordSignupFields: 'USERNAME_AND_EMAIL',
-});
