+++ lib/accounts-config.js
@@ -1,0 +1,6 @@
+
+if(Meteor.isClient){
+	Accounts.ui.config({
+	    passwordSignupFields: 'USERNAME_AND_EMAIL',
+	});
+}
