--- server/main.js
@@ -1,8 +1,0 @@
-import { Meteor } from 'meteor/meteor';
-import { Mongo } from 'meteor/mongo';
-
-Meteor.startup(() => {
-  // code to run on server at startup
-});
-
-import '../imports/collections.js';
